Ref: Split NewDerivationfactory into 2 factories. Change public key serialization to compressed point representation.
