+++ KeyDerivationLibTests/NewTestData.cs
@@ -67,23 +67,23 @@ namespace KeyDerivationLibTests
 
         public static (byte[] scalar, byte[] chainCode) PrivateDerivationKey1 = (
             new byte[] {
-                0xcc, 0xf6, 0xa0, 0xde, 0x4d, 0xc9, 0xcb, 0x52, 0xf6, 0xdf, 0xda, 0x87, 0xe1, 0x6d, 0x54, 0xaa,
-                0x99, 0x7c, 0x16, 0x00, 0x39, 0xd3, 0xcf, 0xfa, 0x19, 0x6a, 0xf6, 0xd5, 0xd2, 0xb9, 0xce, 0xb9
+                0x33, 0xdb, 0xb0, 0x1f, 0x4d, 0xa1, 0xab, 0xcc, 0x29, 0xb8, 0x43, 0x98, 0xed, 0x32, 0xdf, 0xac,
+                0x35, 0x34, 0x0c, 0xf3, 0x2b, 0x12, 0x00, 0x1a, 0x6e, 0x6b, 0x5a, 0xc9, 0x92, 0x96, 0xf7, 0x63
             },
             new byte[] {
-                0xf8, 0x24, 0x90, 0xae, 0xcb, 0x07, 0x81, 0x12, 0xbd, 0x20, 0xb5, 0x02, 0x05, 0x4a, 0x7c, 0x2c,
-                0xee, 0x42, 0x24, 0xaf, 0xb5, 0x92, 0x8f, 0x77, 0x2a, 0xbf, 0x0f, 0x78, 0xc0, 0x2e, 0x9d, 0x74
+                0x43, 0xba, 0x8e, 0x76, 0x69, 0x5d, 0xa7, 0x3b, 0x72, 0x9a, 0x2f, 0x8d, 0xb3, 0x97, 0x94, 0x2c,
+                0x77, 0x9f, 0x94, 0x09, 0xb2, 0x93, 0x38, 0xbc, 0x43, 0x48, 0xa5, 0x70, 0x9a, 0x0f, 0x2e, 0xad
             }
         );
 
         public static (byte[] scalar, byte[] chainCode) PrivateDerivationKey2 = (
             new byte[] {
-                0x5f, 0x9a, 0xe0, 0x60, 0x8d, 0xf8, 0xde, 0x00, 0xf7, 0x1c, 0x9b, 0x9c, 0x38, 0x65, 0x67, 0x5e,
-                0x96, 0xc0, 0x2d, 0x8e, 0xf0, 0x29, 0xf6, 0xd5, 0xaa, 0xfc, 0x4c, 0xae, 0xb2, 0x47, 0x88, 0xd1
+                0x85, 0xfd, 0x5b, 0x88, 0x7e, 0xc9, 0xa8, 0x2a, 0xd7, 0x97, 0x2a, 0xc8, 0x3f, 0xce, 0x39, 0xd4,
+                0xe0, 0x70, 0xc0, 0xed, 0x46, 0x50, 0x3b, 0x84, 0xe0, 0x9a, 0x3f, 0xbf, 0xc6, 0xe7, 0x8e, 0x77
             },
             new byte[] {
-                0x7b, 0x85, 0x52, 0x37, 0x00, 0xb7, 0x87, 0x74, 0x38, 0xb6, 0x59, 0x14, 0xad, 0x22, 0x00, 0x57,
-                0xab, 0x71, 0x6d, 0xa1, 0x82, 0x05, 0x2a, 0x7e, 0x22, 0xb8, 0x90, 0xb2, 0x35, 0x54, 0xd4, 0x88
+                0xfe, 0x1a, 0x7a, 0x0e, 0x1e, 0xf5, 0x09, 0x7c, 0x97, 0x97, 0xe7, 0x4e, 0xf7, 0xfa, 0xf6, 0xef,
+                0x7a, 0x05, 0x7b, 0x9d, 0xc7, 0x13, 0x21, 0xdf, 0x6f, 0x42, 0xfc, 0xd9, 0x96, 0x84, 0x4f, 0x99
             }
         );
 
@@ -92,14 +92,14 @@ namespace KeyDerivationLibTests
 
         public static readonly byte[] PrivateChildKey1 = new byte[32]
         {
-            0xc3, 0xe3, 0x5f, 0xe4, 0xd0, 0x66, 0x3b, 0x06, 0xbf, 0x03, 0x2b, 0x3d, 0xe2, 0x10, 0x34, 0xc1,
-            0x99, 0xc3, 0x3c, 0x12, 0xc8, 0xd0, 0x6c, 0xae, 0x0a, 0xb4, 0x41, 0x98, 0x80, 0xce, 0x06, 0x38
+            0x2a, 0xb8, 0xeb, 0x1b, 0x29, 0xa2, 0xaf, 0x95, 0xe8, 0x43, 0x26, 0xab, 0x9e, 0xe3, 0xb7, 0x47,
+            0xa7, 0x7c, 0x0f, 0xbe, 0x08, 0xdd, 0x3a, 0x7f, 0x44, 0x32, 0x30, 0xf8, 0x9a, 0xef, 0x01, 0x0b
         };
 
         public static readonly byte[] PrivateChildKey2 = new byte[32]
         {
-            0xb7, 0x48, 0x94, 0xbf, 0x2d, 0xd3, 0xd9, 0x68, 0x7d, 0xf0, 0x6b, 0x9f, 0x56, 0x85, 0x5a, 0x26,
-            0xbe, 0x16, 0x45, 0xbf, 0x72, 0x8e, 0x2d, 0xe1, 0x35, 0xd2, 0x56, 0x17, 0x80, 0xdb, 0xbd, 0xe7
+            0x5d, 0x2b, 0x28, 0x10, 0x15, 0x69, 0x0d, 0x12, 0x2a, 0x5e, 0x38, 0xf7, 0x81, 0xc7, 0x8b, 0x42,
+            0xc7, 0x8b, 0x84, 0x01, 0x77, 0x14, 0x0b, 0xfe, 0x70, 0xc3, 0xa8, 0x4c, 0xf3, 0xe0, 0x5a, 0x12
         };
 
         public const string RightTag = "test@user.net";
@@ -109,24 +109,24 @@ namespace KeyDerivationLibTests
         public static (byte[] compressedKey, byte[] chainCode) InitailPublicKey = (
            new byte[] {
                 0x02, 
-                0x63, 0xc1, 0x8e, 0xf9, 0xd2, 0xcb, 0xd4, 0x1a, 0xb2, 0x9f, 0xa2, 0xf8, 0xa6, 0x3b, 0xe4, 0x64,
-                0x88, 0x00, 0xa9, 0x62, 0x47, 0xa7, 0xde, 0x26, 0x8b, 0xc8, 0x1e, 0x76, 0xf3, 0x71, 0x5c, 0x22
+                0x5a, 0xe9, 0x86, 0xab, 0xa7, 0xed, 0xf6, 0x92, 0x46, 0xd5, 0x1f, 0x95, 0x40, 0xe0, 0xf8, 0xb0,
+                0x14, 0x7e, 0x73, 0xd1, 0x55, 0xda, 0xaa, 0x0b, 0x2b, 0xf4, 0xc2, 0x2a, 0xcb, 0x41, 0x00, 0x6e
            },
            new byte[] {
-                0xf8, 0x24, 0x90, 0xae, 0xcb, 0x07, 0x81, 0x12, 0xbd, 0x20, 0xb5, 0x02, 0x05, 0x4a, 0x7c, 0x2c,
-                0xee, 0x42, 0x24, 0xaf, 0xb5, 0x92, 0x8f, 0x77, 0x2a, 0xbf, 0x0f, 0x78, 0xc0, 0x2e, 0x9d, 0x74
+                0x43, 0xba, 0x8e, 0x76, 0x69, 0x5d, 0xa7, 0x3b, 0x72, 0x9a, 0x2f, 0x8d, 0xb3, 0x97, 0x94, 0x2c,
+                0x77, 0x9f, 0x94, 0x09, 0xb2, 0x93, 0x38, 0xbc, 0x43, 0x48, 0xa5, 0x70, 0x9a, 0x0f, 0x2e, 0xad
            }
        );
 
         public static (byte[] compressedKey, byte[] chainCode) ResultPublicKey = (
             new byte[] {
-                0x03, 
-                0x61, 0xf1, 0x0a, 0x0b, 0xec, 0x46, 0xc7, 0x49, 0x26, 0x44, 0xb9, 0xde, 0xac, 0x0d, 0x87, 0x94,
-                0x39, 0x90, 0xfa, 0x8c, 0xc5, 0x36, 0x4d, 0xcc, 0x61, 0x98, 0x76, 0x81, 0x70, 0x26, 0x1b, 0xf3
+                0x02, 
+                0xa4, 0xcd, 0x67, 0xde, 0x67, 0xff, 0xc4, 0x73, 0x4d, 0xfe, 0xd6, 0x7b, 0xde, 0x48, 0xfb, 0x93,
+                0x25, 0x6c, 0x87, 0x70, 0x7d, 0x7c, 0x5e, 0xa1, 0x4f, 0x44, 0xef, 0x77, 0xc2, 0x7b, 0xfc, 0x03
             },
             new byte[] {
-                0x37, 0x59, 0x66, 0x83, 0x01, 0x7b, 0x91, 0x86, 0x31, 0x4b, 0x20, 0x24, 0x3b, 0xb3, 0xbd, 0xfe,
-                0xf3, 0xd4, 0x6e, 0xd0, 0xff, 0xf6, 0xfe, 0xbc, 0x28, 0xd7, 0x7e, 0x73, 0xe5, 0xf2, 0x70, 0x70
+                0x84, 0xd6, 0xd3, 0x9c, 0xc1, 0xb3, 0xbb, 0xe6, 0x6d, 0xeb, 0xed, 0x8f, 0xbe, 0xe1, 0x48, 0xdd,
+                0xad, 0x0e, 0xd5, 0xd7, 0xbc, 0x2f, 0x8c, 0xb6, 0xf5, 0x89, 0x9d, 0x8e, 0x27, 0xaf, 0x81, 0xfd
             }
         );
     }
